refactor: defer integer literal parsing

+++ crates/compiler/src/anf.rs
@@ -20,7 +20,7 @@ pub enum ImmExpr {
     ImmVar { name: String, ty: Ty },
     ImmUnit { ty: Ty },
     ImmBool { value: bool, ty: Ty },
-    ImmInt { value: i64, ty: Ty },
+    ImmInt { value: i128, ty: Ty },
     ImmFloat { value: f64, ty: Ty },
     ImmString { value: String, ty: Ty },
     ImmTag { index: usize, ty: Ty },
